# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/": {
            "get": {
                "produces": [
                    "text/html"
                ],
                "tags": [
                    "Static"
                ],
                "summary": "Metric list",
                "operationId": "printAllMetricStatic",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/ping": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "summary": "Ping",
                "operationId": "pingGetJSON",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update metric value",
                "operationId": "updateMetricPostJSON",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/update/counter/{statName}/{statValue}": {
            "post": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update counter metric",
                "operationId": "updateCounterPost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "statName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Значение",
                        "name": "statValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update/gauge/{statName}/{statValue}": {
            "post": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update gauge metric",
                "operationId": "updateGaugePost",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "statName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Значение",
                        "name": "statValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/update/{statType}/{statName}/{statValue}": {
            "post": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update not implemented",
                "operationId": "updateNotImplementedPost",
                "parameters": [
                    {
                        "enum": [
                            "gauge",
                            "counter"
                        ],
                        "type": "string",
                        "default": "gauge",
                        "description": "Тип метрики",
                        "name": "statType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "statName",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Значение",
                        "name": "statValue",
                        "in": "query"
                    }
                ],
                "responses": {
                    "501": {
                        "description": "Not Implemented"
                    }
                }
            }
        },
        "/updates/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Update"
                ],
                "summary": "Update metric value using batch JSON",
                "operationId": "updateMetricBatchJSON",
                "parameters": [
                    {
                        "description": "JSON",
                        "name": "JSON",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/storage.Metric"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                }
            }
        },
        "/value/": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Value"
                ],
                "summary": "Metric value JSON",
                "operationId": "metricValuePostJSON",
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/value/{statType}/{statName}": {
            "get": {
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Value"
                ],
                "summary": "Metric value",
                "operationId": "printMetricGet",
                "parameters": [
                    {
                        "enum": [
                            "gauge",
                            "counter"
                        ],
                        "type": "string",
                        "default": "gauge",
                        "description": "Тип метрики",
                        "name": "statType",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "Имя метрики",
                        "name": "statName",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        }
    },
    "definitions": {
        "storage.Metric": {
            "type": "object",
            "properties": {
                "delta": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "number"
                }
            }
        }
    }
}